refactor: improve targets handling
