+++ src/transit.py
@@ -12,16 +12,14 @@ from src.astro import CelestialObject
 from src.constants import (
     ASTRO_EPHEMERIS,
     CHANGE_ELEVATION,
-    EARTH_RADIOUS,
     FLIGHTS_SEARCH_URL,
     INTERVAL_IN_SECS,
     NUM_SECONDS_PER_MIN,
-    TEST_DATA_PATH,
     TOP_MINUTE,
-    Altitude,
     PossibilityLevel,
 )
-from src.flight_data import get_flight_data, load_existing_flight_data, parse_fligh_data
+from src.demo import generate_test_flightaware_data
+from src.flight_data import get_flight_data, parse_fligh_data
 from src.position import (
     AreaBoundingBox,
     geographic_to_altaz,
@@ -118,12 +116,9 @@ def check_transit(
     flight: dict,
     window_time: list,
     ref_datetime: datetime,
-    my_position: Topos,
+    observer_position: Topos,
     target: CelestialObject,
     earth_ref,
-    test_mode: bool = False,
-    observer_lat: float = 0.0,
-    observer_lon: float = 0.0,
 ) -> dict:
     """Given the data of a flight, compute a possible transit with the target.
 
@@ -137,15 +132,13 @@ def check_transit(
     ref_datetime: datetime
         Reference datetime, deltas from window_time will be add to this reference to compute the future position
         of plane and target.
-    my_position: Topos
+    observer_position: Topos
         Object from skifield library which was instanced with current position of the observer (
         latitude, longitude and elevation).
     target: CelestialObject
         It could be the Moon or Sun, or whatever celestial object to compute a possible transit.
     earth_ref: Any
         Earth data gotten from the de421.bsp database by NASA's JPL.
-    test_mode: bool
-        If True, evaluates current position (t=0) for static test aircraft.
 
     Returns
     -------
@@ -160,49 +153,13 @@ def check_transit(
     update_response = False
     POSSIBLE_TRANSIT_LEVELS = {PossibilityLevel.HIGH.value, PossibilityLevel.MEDIUM.value}
 
-    # Calculate horizontal distance from observer to aircraft in nautical miles
+    # Calculate horizontal distance from observer to aircraft in kilometers
     distance_km = haversine_distance(
-        observer_lat, observer_lon, flight["latitude"], flight["longitude"]
+        float(observer_position.target.latitude.degrees),
+        float(observer_position.target.longitude.degrees),
+        flight["latitude"],
+        flight["longitude"]
     )
-
-    # # In test mode, check current position first (t=0, static aircraft)
-    # if test_mode:
-    #     alt_diff = abs(current_alt - target.altitude.degrees)
-    #     az_diff = abs(current_az - target.azimuthal.degrees)
-    #     angular_sep = calculate_angular_separation(alt_diff, az_diff)
-
-    #     min_angular_sep = angular_sep
-
-    #     # Always record if aircraft is above horizon, regardless of separation
-    #     if current_alt > 0:
-    #         response = {
-    #             "id": flight["name"],
-    #             "aircraft_type": flight.get("aircraft_type", "N/A"),
-    #             "fa_flight_id": flight.get("fa_flight_id", ""),
-    #             "origin": flight["origin"],
-    #             "destination": flight["destination"],
-    #             "alt_diff": round(float(alt_diff), 3),
-    #             "az_diff": round(float(az_diff), 3),
-    #             "angular_separation": round(float(angular_sep), 3),
-    #             "time": 0.0,  # Current position
-    #             "target_alt": initial_target_alt,
-    #             "plane_alt": round(float(current_alt), 2),
-    #             "target_az": initial_target_az,
-    #             "plane_az": round(float(current_az), 2),
-    #             "is_possible_transit": 1 if angular_sep <= 6.0 else 0,
-    #             "possibility_level": get_possibility_level(angular_sep),
-    #             "elevation_change": CHANGE_ELEVATION.get(
-    #                 flight["elevation_change"], None
-    #             ),
-    #             "direction": flight["direction"],
-    #             "speed": flight["speed"],
-    #             "target": target.name,
-    #             "latitude": flight["latitude"],
-    #             "longitude": flight["longitude"],
-    #             "aircraft_elevation": flight.get("elevation", 0),  # Actual altitude in meters
-    #             "aircraft_elevation_feet": flight.get("elevation_feet", 0),  # Actual altitude in feet
-    #             "distance_km": round(distance_km, 1),  # Distance from observer in km
-    #         }
 
     for idx, minute in enumerate(window_time):
         # Get future position of plane
@@ -222,7 +179,7 @@ def check_transit(
             future_lon,
             flight["elevation"],
             earth_ref,
-            my_position,
+            observer_position,
             future_time,
         )
 
@@ -292,338 +249,6 @@ def check_transit(
     return response
 
 
-def generate_mock_results(obs_lat: float, obs_lon: float, obs_elev: float) -> dict:
-    """Generate mock transit results for demonstration purposes.
-
-    Returns hardcoded results showing HIGH, MEDIUM, LOW, and NONE classifications
-    for both moon and sun targets.
-    """
-    # Fixed celestial target positions
-    moon_az, moon_alt = 135.0, 40.0
-    sun_az, sun_alt = 225.0, 35.0
-
-    # Helper to create aircraft position at specific azimuth and distance
-    # Uses haversine formula to match the map's azimuth arrow calculation
-    def position_at(azimuth_deg, distance_km):
-        import math
-        d = distance_km / EARTH_RADIOUS  # Angular distance in radians
-
-        brng = math.radians(azimuth_deg)
-        lat1 = math.radians(obs_lat)
-        lon1 = math.radians(obs_lon)
-
-        lat2 = math.asin(
-            math.sin(lat1) * math.cos(d) +
-            math.cos(lat1) * math.sin(d) * math.cos(brng)
-        )
-
-        lon2 = lon1 + math.atan2(
-            math.sin(brng) * math.sin(d) * math.cos(lat1),
-            math.cos(d) - math.sin(lat1) * math.sin(lat2)
-        )
-
-        return round(math.degrees(lat2), 6), round(math.degrees(lon2), 6)
-
-    flights = []
-
-    # MOON TRANSITS
-    # HIGH - nearly perfect alignment (≤1°)
-    lat, lon = position_at(moon_az, 15)  # 15 km on moon bearing
-    alt_diff, az_diff = 0.5, 0.3
-    flights.append({
-        "id": "MOON_HIGH",
-        "aircraft_type": "A320",
-        "fa_flight_id": "MOON_HIGH-test-123",
-        "origin": "Los Angeles",
-        "destination": "San Diego",
-        "alt_diff": alt_diff,
-        "az_diff": az_diff,
-        "angular_separation": round(np.sqrt(alt_diff**2 + az_diff**2), 3),
-        "time": 2.5,
-        "target_alt": moon_alt,
-        "plane_alt": 40.5,
-        "target_az": moon_az,
-        "plane_az": 135.3,
-        "is_possible_transit": 1,
-        "possibility_level": 3,  # HIGH
-        "elevation_change": "descending",
-        "direction": 315,
-        "target": "moon",
-        "latitude": lat,
-        "longitude": lon,
-        "aircraft_elevation": 10668,  # 35,000 ft in meters
-        "aircraft_elevation_feet": 35000,  # 35,000 ft
-        "distance_km": 15,  # 15 km = 8.1 nm from observer
-    })
-
-    # MEDIUM - moderate alignment (≤2°)
-    lat, lon = position_at(moon_az - 2, 20)  # 20 km, offset 2° from moon bearing
-    alt_diff, az_diff = 1.2, 1.0
-    flights.append({
-        "id": "MOON_MED",
-        "aircraft_type": "B737",
-        "fa_flight_id": "MOON_MED-test-456",
-        "origin": "Phoenix",
-        "destination": "San Diego",
-        "alt_diff": alt_diff,
-        "az_diff": az_diff,
-        "angular_separation": round(np.sqrt(alt_diff**2 + az_diff**2), 3),
-        "time": 3.2,
-        "target_alt": moon_alt,
-        "plane_alt": 38.8,
-        "target_az": moon_az,
-        "plane_az": 134.0,
-        "is_possible_transit": 1,
-        "possibility_level": 2,  # MEDIUM
-        "elevation_change": "descending",
-        "direction": 310,
-        "target": "moon",
-        "latitude": lat,
-        "longitude": lon,
-        "aircraft_elevation": 10972,  # 36,000 ft in meters
-        "aircraft_elevation_feet": 36000,  # 36,000 ft
-        "distance_km": 20,  # 20 km = 10.8 nm from observer
-    })
-
-    # LOW - marginal alignment (≤6°)
-    lat, lon = position_at(moon_az + 7, 25)  # 25 km, offset 7° from moon bearing
-    alt_diff, az_diff = 4.0, 3.5
-    flights.append({
-        "id": "MOON_LOW",
-        "aircraft_type": "A321",
-        "fa_flight_id": "MOON_LOW-test-789",
-        "origin": "San Francisco",
-        "destination": "San Diego",
-        "alt_diff": alt_diff,
-        "az_diff": az_diff,
-        "angular_separation": round(np.sqrt(alt_diff**2 + az_diff**2), 3),
-        "time": 4.8,
-        "target_alt": moon_alt,
-        "plane_alt": 36.0,
-        "target_az": moon_az,
-        "plane_az": 138.5,
-        "is_possible_transit": 1,
-        "possibility_level": 1,  # LOW
-        "elevation_change": "descending",
-        "direction": 305,
-        "target": "moon",
-        "latitude": lat,
-        "longitude": lon,
-        "aircraft_elevation": 11277,  # 37,000 ft in meters
-        "aircraft_elevation_feet": 37000,  # 37,000 ft
-        "distance_km": 25,  # 25 km = 13.5 nm from observer
-    })
-
-    # SUN TRANSITS
-    # HIGH - nearly perfect alignment (≤1°)
-    lat, lon = position_at(sun_az, 15)  # 15 km on sun bearing
-    alt_diff, az_diff = 0.4, 0.6
-    flights.append({
-        "id": "SUN_HIGH",
-        "aircraft_type": "B777",
-        "fa_flight_id": "SUN_HIGH-test-111",
-        "origin": "Las Vegas",
-        "destination": "San Diego",
-        "alt_diff": alt_diff,
-        "az_diff": az_diff,
-        "angular_separation": round(np.sqrt(alt_diff**2 + az_diff**2), 3),
-        "time": 2.8,
-        "target_alt": sun_alt,
-        "plane_alt": 35.4,
-        "target_az": sun_az,
-        "plane_az": 225.6,
-        "is_possible_transit": 1,
-        "possibility_level": 3,  # HIGH
-        "elevation_change": "descending",
-        "direction": 45,
-        "target": "sun",
-        "latitude": lat,
-        "longitude": lon,
-        "aircraft_elevation": 10363,  # 34,000 ft in meters
-        "aircraft_elevation_feet": 34000,  # 34,000 ft
-        "distance_km": 15,  # 15 km = 8.1 nm from observer
-    })
-
-    # MEDIUM - moderate alignment (≤2°)
-    lat, lon = position_at(sun_az + 2, 20)  # 20 km, offset 2° from sun bearing
-    alt_diff, az_diff = 1.3, 1.1
-    flights.append({
-        "id": "SUN_MED",
-        "aircraft_type": "A330",
-        "fa_flight_id": "SUN_MED-test-222",
-        "origin": "Denver",
-        "destination": "San Diego",
-        "alt_diff": alt_diff,
-        "az_diff": az_diff,
-        "angular_separation": round(np.sqrt(alt_diff**2 + az_diff**2), 3),
-        "time": 3.5,
-        "target_alt": sun_alt,
-        "plane_alt": 33.7,
-        "target_az": sun_az,
-        "plane_az": 226.1,
-        "is_possible_transit": 1,
-        "possibility_level": 2,  # MEDIUM
-        "elevation_change": "descending",
-        "direction": 40,
-        "target": "sun",
-        "latitude": lat,
-        "longitude": lon,
-        "aircraft_elevation": 10058,  # 33,000 ft in meters
-        "aircraft_elevation_feet": 33000,  # 33,000 ft
-        "distance_km": 20,  # 20 km = 10.8 nm from observer
-    })
-
-    # LOW - marginal alignment (≤6°)
-    lat, lon = position_at(sun_az - 7, 25)  # 25 km, offset 7° from sun bearing
-    alt_diff, az_diff = 3.8, 4.2
-    flights.append({
-        "id": "SUN_LOW",
-        "aircraft_type": "B787",
-        "fa_flight_id": "SUN_LOW-test-333",
-        "origin": "Oakland",
-        "destination": "San Diego",
-        "alt_diff": alt_diff,
-        "az_diff": az_diff,
-        "angular_separation": round(np.sqrt(alt_diff**2 + az_diff**2), 3),
-        "time": 5.2,
-        "target_alt": sun_alt,
-        "plane_alt": 31.2,
-        "target_az": sun_az,
-        "plane_az": 220.8,
-        "is_possible_transit": 1,
-        "possibility_level": 1,  # LOW
-        "elevation_change": "descending",
-        "direction": 35,
-        "target": "sun",
-        "latitude": lat,
-        "longitude": lon,
-        "aircraft_elevation": 9754,  # 32,000 ft in meters
-        "aircraft_elevation_feet": 32000,  # 32,000 ft
-        "distance_km": 25,  # 25 km = 13.5 nm from observer
-    })
-
-    # UNLIKELY - no transit (far from both targets, >6°)
-    lat, lon = position_at(0, 25)  # North, 25 km
-    # This plane is heading North (0°), far from moon at 135°
-    plane_alt_1, plane_az_1 = 25.0, 5.0  # Low on horizon, heading north
-    alt_diff_1 = abs(plane_alt_1 - moon_alt)  # 15°
-    az_diff_1 = abs(plane_az_1 - moon_az)  # 130°
-    flights.append({
-        "id": "NONE_01",
-        "aircraft_type": "B737",
-        "fa_flight_id": "NONE_01-test-444",
-        "origin": "San Diego",
-        "destination": "San Francisco",
-        "alt_diff": round(alt_diff_1, 3),
-        "az_diff": round(az_diff_1, 3),
-        "angular_separation": round(np.sqrt(alt_diff_1**2 + az_diff_1**2), 3),
-        "time": None,
-        "target_alt": moon_alt,
-        "plane_alt": plane_alt_1,
-        "target_az": moon_az,
-        "plane_az": plane_az_1,
-        "is_possible_transit": 0,
-        "possibility_level": 0,  # UNLIKELY
-        "elevation_change": "climbing",
-        "direction": 0,
-        "target": "moon",
-        "latitude": lat,
-        "longitude": lon,
-        "aircraft_elevation": 7620,  # 25,000 ft in meters
-        "aircraft_elevation_feet": 25000,  # 25,000 ft
-        "distance_km": 25,  # 25 km = 13.5 nm from observer
-    })
-
-    lat, lon = position_at(180, 25)  # South, 25 km
-    # This plane is heading South (180°), somewhat close to sun at 225°
-    plane_alt_2, plane_az_2 = 32.0, 185.0  # Mid-altitude, heading south
-    alt_diff_2 = abs(plane_alt_2 - sun_alt)  # 3°
-    az_diff_2 = abs(plane_az_2 - sun_az)  # 40°
-    flights.append({
-        "id": "NONE_02",
-        "aircraft_type": "A320",
-        "fa_flight_id": "NONE_02-test-555",
-        "origin": "San Diego",
-        "destination": "Denver",
-        "alt_diff": round(alt_diff_2, 3),
-        "az_diff": round(az_diff_2, 3),
-        "angular_separation": round(np.sqrt(alt_diff_2**2 + az_diff_2**2), 3),
-        "time": None,
-        "target_alt": sun_alt,
-        "plane_alt": plane_alt_2,
-        "target_az": sun_az,
-        "plane_az": plane_az_2,
-        "is_possible_transit": 0,
-        "possibility_level": 0,  # UNLIKELY
-        "elevation_change": "level",
-        "direction": 180,
-        "target": "sun",
-        "latitude": lat,
-        "longitude": lon,
-        "aircraft_elevation": 9144,  # 30,000 ft in meters
-        "aircraft_elevation_feet": 30000,  # 30,000 ft
-        "distance_km": 25,  # 25 km = 13.5 nm from observer
-    })
-
-    lat, lon = position_at(270, 25)  # West, 25 km
-    # This plane is heading West (270°), far from moon at 135°
-    plane_alt_3, plane_az_3 = 15.0, 275.0  # Low altitude private plane, heading west
-    alt_diff_3 = abs(plane_alt_3 - moon_alt)  # 25°
-    az_diff_3 = abs(plane_az_3 - moon_az)  # 140°
-    flights.append({
-        "id": "PRIV01",
-        "aircraft_type": "SR22",
-        "fa_flight_id": "PRIV01-test-666",
-        "origin": "San Diego",
-        "destination": "N/D",
-        "alt_diff": round(alt_diff_3, 3),
-        "az_diff": round(az_diff_3, 3),
-        "angular_separation": round(np.sqrt(alt_diff_3**2 + az_diff_3**2), 3),
-        "time": None,
-        "target_alt": moon_alt,
-        "plane_alt": plane_alt_3,
-        "target_az": moon_az,
-        "plane_az": plane_az_3,
-        "is_possible_transit": 0,
-        "possibility_level": 0,  # UNLIKELY
-        "elevation_change": "level",
-        "direction": 270,
-        "target": "moon",
-        "latitude": lat,
-        "longitude": lon,
-        "aircraft_elevation": 1524,  # 5,000 ft in meters (private plane)
-        "aircraft_elevation_feet": 5000,  # 5,000 ft
-        "distance_km": 25,  # 25 km = 13.5 nm from observer
-    })
-
-    return {
-        "flights": flights,
-        "targetCoordinates": {
-            "moon": {"altitude": moon_alt, "azimuthal": moon_az},
-            "sun": {"altitude": sun_alt, "azimuthal": sun_az}
-        },
-        "trackingTargets": ["moon", "sun"],
-        "weather": {
-            "cloud_cover": 0,
-            "condition": "clear",
-            "icon": "☀️",
-            "description": "clear sky",
-            "api_success": True
-        },
-        "boundingBox": {
-            "latLowerLeft": obs_lat - 0.5,
-            "lonLowerLeft": obs_lon - 0.5,
-            "latUpperRight": obs_lat + 0.5,
-            "lonUpperRight": obs_lon + 0.5,
-        },
-        "observerPosition": {
-            "latitude": obs_lat,
-            "longitude": obs_lon,
-            "elevation": obs_elev,
-        },
-    }
-
-
 def get_transits(
     latitude: float,
     longitude: float,
@@ -646,94 +271,59 @@ def get_transits(
     custom_bbox : dict
         Optional custom bounding box with keys: lat_lower_left, lon_lower_left, lat_upper_right, lon_upper_right
     """
-    # MOCK MODE - return hardcoded demo results
-    if test_mode:
-        logger.info("🎭 MOCK MODE: Returning demonstration results")
-        return generate_mock_results(latitude, longitude, elevation)
     API_KEY = os.getenv("AEROAPI_API_KEY")
     WEATHER_API_KEY = os.getenv("OPENWEATHER_API_KEY")
     MIN_ALTITUDE = min_altitude if min_altitude is not None else float(os.getenv("MIN_TARGET_ALTITUDE", 15))
-
-    logger.info(f"{latitude=}, {longitude=}, {elevation=}, {target_name=}")
-
-    # Check weather conditions
-    is_clear, weather_info = get_weather_condition(latitude, longitude, WEATHER_API_KEY)
-    logger.info(f"Weather check: clear={is_clear}, {weather_info}")
-    
-    MY_POSITION = get_my_pos(
+    OBSERVER_POSITION = get_my_pos(
         lat=latitude,
         lon=longitude,
         elevation=elevation,
         base_ref=EARTH,
     )
 
+    logger.info(f"{latitude=}, {longitude=}, {elevation=}, {target_name=}")
+
+    # MOCK MODE - return hardcoded demo results
+    # if test_mode:
+    #     logger.info("🎭 MOCK MODE: Returning demonstration results")
+    #     return generate_demo_flight_data(latitude, longitude, elevation, target_name)
+
+    # Check weather conditions
+    is_clear, weather_info = get_weather_condition(latitude, longitude, WEATHER_API_KEY, test_mode)
+    logger.info(f"Weather check: clear={is_clear}, {weather_info}")
+
     window_time = np.linspace(
         0, TOP_MINUTE, TOP_MINUTE * (NUM_SECONDS_PER_MIN // INTERVAL_IN_SECS)
     )
     logger.info(f"number of times to check for each flight: {len(window_time)}")
-    
+
     # Get the local timezone using tzlocal
     local_timezone = get_localzone_name()
     naive_datetime_now = datetime.now()
     ref_datetime = naive_datetime_now.replace(tzinfo=ZoneInfo(local_timezone))
 
     # Determine which targets to check
+    target_names = ["moon", "sun"] if target_name == "auto" else [target_name]
     targets_to_check = []
     target_coordinates = {}
 
-    # In test mode, use fake positions from test data metadata
-    test_overrides = {}
-    if test_mode:
-        try:
-            test_data = load_existing_flight_data(TEST_DATA_PATH)
-            meta = test_data.get("_test_metadata", {})
-            test_overrides = {
-                "moon": {
-                    "altitude": meta.get("moon_altitude", 60),
-                    "azimuth": meta.get("moon_azimuth", 180),
-                },
-                "sun": {
-                    "altitude": meta.get("sun_altitude", 55),
-                    "azimuth": meta.get("sun_azimuth", 200),
-                },
-            }
-            logger.info(f"Test mode: using fake positions {test_overrides}")
-        except Exception:
-            pass
-
-    if target_name == "auto":
-        # Check both moon and sun if conditions permit
-        for target in ["moon", "sun"]:
-            overrides = test_overrides.get(target) if test_mode else None
-            obj = CelestialObject(name=target, observer_position=MY_POSITION, test_overrides=overrides)
-            obj.update_position(ref_datetime=ref_datetime)
-            coords = obj.get_coordinates()
-
-            target_coordinates[target] = coords
-
-            if coords["altitude"] >= MIN_ALTITUDE and is_clear:
-                targets_to_check.append(target)
-                logger.info(f"{target} at {coords['altitude']}° az {coords['azimuthal']}° - tracking enabled")
-            else:
-                reason = "below horizon" if coords["altitude"] < MIN_ALTITUDE else "weather"
-                logger.info(f"{target} at {coords['altitude']}° - skipped ({reason})")
-    else:
-        # Single target mode
-        overrides = test_overrides.get(target_name) if test_mode else None
-        obj = CelestialObject(name=target_name, observer_position=MY_POSITION, test_overrides=overrides)
+    # Check both moon and sun if conditions permit
+    for target in target_names:
+        obj = CelestialObject(name=target, observer_position=OBSERVER_POSITION)
         obj.update_position(ref_datetime=ref_datetime)
         coords = obj.get_coordinates()
 
-        target_coordinates[target_name] = coords
+        target_coordinates[target] = coords
 
         if coords["altitude"] >= MIN_ALTITUDE and is_clear:
-            targets_to_check.append(target_name)
+            targets_to_check.append(target)
+            logger.info(f"{target} at {coords['altitude']}° az {coords['azimuthal']}° - tracking enabled")
         else:
             reason = "below horizon" if coords["altitude"] < MIN_ALTITUDE else "weather"
-            logger.warning(f"{target_name} not trackable ({reason})")
+            logger.info(f"{target} at {coords['altitude']}° - skipped ({reason})")
 
     data = list()
-    tracking_targets = targets_to_check.copy()  # For response
+    tracking_targets = targets_to_check.copy() # For response
 
     # Use custom bounding box if provided, otherwise use default
     if custom_bbox:
@@ -750,8 +340,8 @@ def get_transits(
     if targets_to_check:
         # Fetch flight data once
         if test_mode:
-            raw_flight_data = load_existing_flight_data(TEST_DATA_PATH)
-            logger.info("Loading existing flight data since is using TEST mode")
+            logger.info("🧪 TEST MODE: generating test flight data...")
+            raw_flight_data = generate_test_flightaware_data(OBSERVER_POSITION, targets_to_check, target_coordinates)
         else:
             raw_flight_data = get_flight_data(search_bbox, FLIGHTS_SEARCH_URL, API_KEY)
 
@@ -763,9 +353,8 @@ def get_transits(
 
         # Check transits for each target
         for target in targets_to_check:
-            overrides = test_overrides.get(target) if test_mode else None
-            celestial_obj = CelestialObject(name=target, observer_position=MY_POSITION, test_overrides=overrides)
-            celestial_obj.update_position(ref_datetime=ref_datetime)
+            celestial_obj = CelestialObject(name=target, observer_position=OBSERVER_POSITION)
+            #celestial_obj.update_position(ref_datetime=ref_datetime)
 
             for flight in flight_data:
                 celestial_obj.update_position(ref_datetime=ref_datetime)
@@ -774,12 +363,9 @@ def get_transits(
                     flight,
                     window_time,
                     ref_datetime,
-                    MY_POSITION,
+                    OBSERVER_POSITION,
                     celestial_obj,
                     EARTH,
-                    test_mode=test_mode,
-                    observer_lat=latitude,
-                    observer_lon=longitude,
                 )
                 data.append(transit_result)
                 logger.info(transit_result)
